+++ checkticket_api/checkticket_api/serve.py
@@ -12,7 +12,7 @@ Development: call like
 Several configuration methods are shown within the code.
 
 
-Copyright (C) 2016, 2017 by Bundesamt für Sicherheit in der Informationstechnik
+Copyright (C) 2016-2018 by Bundesamt für Sicherheit in der Informationstechnik
 
 Software engineering by Intevation GmbH
 
@@ -34,15 +34,15 @@ Author(s):
     * Dustin Demuth <dustin@intevation.de>
 """
 
-# The intelmqmail module needs an UTF-8 locale, so we set a common one
-# available in Ubuntu 14.04/LTS here explicitely. This also removes the
-# necessity to configure the calling http server to set the locale correctly.
 import os
-os.environ['LANG']= 'en_US.UTF-8'
-
 from psycopg2.extras import DictConnection
 import hug
 import logging
+
+# The intelmqmail module needs an UTF-8 locale, so we set a common one
+# available in Ubuntu 14.04/LTS here explicitely. This also removes the
+# necessity to configure the calling http server to set the locale correctly.
+os.environ['LANG'] = 'en_US.UTF-8'
 
 log = logging.getLogger(__name__)
 # adding a custom log level for even more details when diagnosing
@@ -66,6 +66,7 @@ config = None
 conn = None
 cur = None
 
+
 @hug.startup()
 def setup(api):
     global config, conn, cur
@@ -76,19 +77,21 @@ def setup(api):
 
 
 @hug.cli()
-@hug.get(ENDPOINT_PREFIX  + '/getEventIDsForTicket')
-def getEventIDsForTicket(ticket:hug.types.length(17, 18)):
+@hug.get(ENDPOINT_PREFIX + '/getEventIDsForTicket')
+def getEventIDsForTicket(ticket: hug.types.length(17, 18)):
     global cur
     event_ids = []
     try:
-        cur.execute("SELECT array_agg(d.events_id) AS a FROM directives AS d "
+        cur.execute("SELECT array_agg(d.events_id ORDER BY d.events_id) AS a"
+                    "   FROM directives AS d "
                     "   JOIN sent ON d.sent_id = sent.id "
                     "   WHERE sent.intelmq_ticket = %s;", (ticket,))
         event_ids = cur.fetchone()["a"]
     finally:
-        cur.connection.commit() # end transaction
+        cur.connection.commit()  # end transaction
 
     return event_ids
+
 
 class ListOfIds(hug.types.Multiple):
     """Only accept a list of numbers."""
@@ -99,8 +102,8 @@ class ListOfIds(hug.types.Multiple):
 
 
 @hug.cli()
-@hug.get(ENDPOINT_PREFIX  + '/getEvents')
-def getEvents(ids:ListOfIds()):
+@hug.get(ENDPOINT_PREFIX + '/getEvents')
+def getEvents(ids: ListOfIds()):
     global cur
     events = []
 
@@ -109,16 +112,18 @@ def getEvents(ids:ListOfIds()):
         rows = cur.fetchall()
         for row in rows:
             # remove None entries from the resulting dict
-            event = {k:v for k,v in row.items() if v != None}
+            event = {k: v for k, v in row.items() if v is not None}
             events.append(event)
     finally:
-        cur.connection.commit() # end transaction
+        cur.connection.commit()  # end transaction
 
     return events
 
+
 @hug.get(ENDPOINT_PREFIX + '/getEventsForTicket')
-def getEventsForTicket(ticket:hug.types.length(17, 18)):
-    return getEvents(getEventIDsForTicket(ticket))
+def getEventsForTicket(ticket: hug.types.length(17, 18), limit: int=None):
+    """Get events for given ticket up to optional `limit`."""
+    return getEvents(getEventIDsForTicket(ticket)[0:limit])
 
 
 @hug.get(ENDPOINT_PREFIX + '/getLastTicketNumber')
@@ -128,9 +133,10 @@ def getLastTicketNumber():
     try:
         last_ticket_number = db.last_ticket_number(cur)
     finally:
-        cur.connection.commit() # end transaction
+        cur.connection.commit()  # end transaction
 
     return last_ticket_number
+
 
 def main():
     # expose only one function to the cli