+++ events_api/events_api/__main__.py
@@ -4,5 +4,5 @@ https://moshez.wordpress.com/2016/06/07/__name__-__main__-considered-harmful/
 """
 if __name__ != '__main__':
         raise ImportError("This module cannot be imported.")
-from events_api import serve
+from events_api import serve  # noqa (prevent pycodestyle throwing E402)
 serve.main()